requirements.txt
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
geopy==2.4.1
psycopg2-binary==2.9.10